# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/users": {
      "get": {
        "operationId": "AdminController_listUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de usuários retornada com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Lista todos os usuários do sistema",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/clubs": {
      "get": {
        "operationId": "AdminController_listClubs",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de clubes retornada com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClubDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Lista todos os clubes do sistema",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/dependants": {
      "get": {
        "operationId": "AdminController_listDependants",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de dependentes retornada com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DependantsListItemView"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Lista todos os dependentes do sistema",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/users/{userId}/roles": {
      "post": {
        "operationId": "AdminController_manageUserRole",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManageUserRoleDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Perfis do usuário atualizados com sucesso."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Gerencia os perfis (roles) de um usuário",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/users/{userId}/family": {
      "get": {
        "operationId": "AdminController_viewUserFamily",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados da família retornados com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Visualiza os detalhes da família de um usuário específico",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/affiliations": {
      "get": {
        "operationId": "AdminController_listAffiliations",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de afiliações retornada com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AffiliationDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Lista todas as afiliações de famílias e seus status",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/clubs/{clubId}/director": {
      "patch": {
        "operationId": "AdminController_changeClubDirector",
        "parameters": [
          {
            "name": "clubId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeClubDirectorDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Diretor do clube alterado com sucesso."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Altera o diretor (proprietário) de um clube",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/enrollments": {
      "get": {
        "operationId": "AdminController_listAllEnrollments",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de matrículas retornada com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnrollmentRequestDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Lista todas as solicitações de matrícula do sistema",
        "tags": [
          "Admin"
        ]
      }
    },
    "/account/user": {
      "post": {
        "operationId": "AccountController_registerUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserInputDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário registrado com sucesso, retorna tokens.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterUserOutputDto"
                }
              }
            }
          },
          "400": {
            "description": "Dados de entrada inválidos ou email/CPF já em uso."
          }
        },
        "summary": "Registra um novo usuário (responsável familiar)",
        "tags": [
          "Conta de Usuário"
        ]
      }
    },
    "/account/profile": {
      "patch": {
        "operationId": "AccountController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Perfil atualizado com sucesso."
          },
          "400": {
            "description": "Dados de entrada inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Atualiza o perfil do usuário autenticado",
        "tags": [
          "Conta de Usuário"
        ]
      }
    },
    "/account/change-password": {
      "post": {
        "operationId": "AccountController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Senha alterada com sucesso."
          },
          "400": {
            "description": "Dados de entrada inválidos ou nova senha igual à antiga."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Altera a senha do usuário autenticado",
        "tags": [
          "Conta de Usuário"
        ]
      }
    },
...
    "/checkout": {
      "post": {
        "operationId": "CheckoutController_createCheckout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pagamento processado ou iniciado com sucesso. Retorna os dados da transação do gateway."
          },
          "400": {
            "description": "Dados de pagamento inválidos."
          },
          "401": {
            "description": "Não autorizado."
          },
          "404": {
            "description": "Usuário ou família não encontrado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Inicia o processo de pagamento da afiliação familiar",
        "tags": [
          "Afiliação e Pagamento"
        ]
      }
    },
    "/dependants": {
      "post": {
        "operationId": "DependantController_add",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddDependantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Dependente adicionado com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DependantDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Adiciona um novo dependente à família do usuário",
        "tags": [
          "Família e Dependentes"
        ]
      },
      "get": {
        "operationId": "DependantController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de dependentes retornada com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DependantDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Lista todos os dependentes da família do usuário",
        "tags": [
          "Família e Dependentes"
        ]
      }
    },
...
    "/enrollments": {
      "post": {
        "operationId": "EnrollmentController_request",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestEnrollmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Solicitação de matrícula criada com sucesso."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Solicita a matrícula de um dependente em um clube",
        "tags": [
          "Matrículas (Responsável)"
        ]
      }
    },
...
  },
  "info": {
    "title": "API NCFCA",
    "description": "Documentação da API para o sistema NCFCA. Esta é a única fonte de verdade para os contratos da API.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Entre com o token JWT",
        "in": "header"
      }
    },
    "schemas": {
...
      "ClubDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único do clube.",
            "format": "uuid",
            "example": "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11"
          },
          "name": {
            "type": "string",
            "description": "Nome oficial do clube.",
            "example": "Clube de Debate Oradores do Amanhã"
          },
          "city": {
            "type": "string",
            "description": "Cidade de localização do clube.",
            "example": "Brasília"
          },
          "state": {
            "type": "string",
            "description": "Estado de localização do clube.",
            "example": "DF"
          },
          "principalId": {
            "type": "string",
            "description": "ID do usuário que é o diretor do clube (principal).",
            "format": "uuid",
            "example": "f1eebc99-9c0b-4ef8-bb6d-6bb9bd380a22"
          },
          "createdAt": {
            "type": "string",
            "description": "Data de criação do clube.",
            "format": "date-time",
            "example": "2023-01-15T10:00:00Z"
          },
          "corum": {
            "type": "number",
            "description": "Quantidade de membros ativos no clube.",
            "example": 25
          }
        },
        "required": [
          "id",
          "name",
          "city",
          "state",
          "principalId",
          "createdAt",
          "corum"
        ]
      },
...
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único do usuário.",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "description": "Email de contato do usuário.",
            "example": "joao.silva@example.com"
          },
          "firstName": {
            "type": "string",
            "description": "Primeiro nome do usuário.",
            "example": "João"
          },
          "lastName": {
            "type": "string",
            "description": "Último nome do usuário.",
            "example": "Silva"
          },
          "phone": {
            "type": "string",
            "description": "Telefone do usuário.",
            "example": "5511987654321"
          },
          "cpf": {
            "type": "string",
            "description": "CPF do usuário.",
            "example": "12345678900"
          },
          "rg": {
            "type": "string",
            "description": "Rg do usuário.",
            "example": "123456"
          },
          "roles": {
            "type": "array",
            "description": "Perfis de acesso do usuário.",
            "example": [
              "SEM_FUNCAO"
            ],
            "items": {
              "type": "string",
              "enum": [
                "DONO_DE_CLUBE",
                "SEM_FUNCAO",
                "ADMIN"
              ]
            }
          },
          "address": {
            "description": "Endereço completo do usuário.",
            "allOf": [
              {
                "$ref": "#/components/schemas/AddressDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "lastName",
          "phone",
          "cpf",
          "rg",
          "roles",
          "address"
        ]
      },
      "DependantsListItemView": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email do dependente",
            "example": "dependente@example.com"
          },
          "name": {
            "type": "string",
            "description": "Nome completo do dependente",
            "example": "João Silva"
          },
          "id": {
            "type": "string",
            "description": "ID único do dependente",
            "format": "uuid"
          }
        },
        "required": [
          "email",
          "name",
          "id"
        ]
      },
      "ManageUserRoleDto": {
        "type": "object",
        "properties": {
          "roles": {
            "type": "array",
            "description": "Lista de perfis (roles) a serem atribuídos ao usuário. Substitui os perfis existentes.",
            "example": [
              "DONO_DE_CLUBE"
            ],
            "items": {
              "type": "string",
              "enum": [
                "DONO_DE_CLUBE",
                "SEM_FUNCAO",
                "ADMIN"
              ]
            }
          }
        },
        "required": [
          "roles"
        ]
      },
...
      "FamilyDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único da família.",
            "format": "uuid",
            "example": "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11"
          },
          "holderId": {
            "type": "string",
            "description": "ID do usuário responsável pela família (holder).",
            "format": "uuid",
            "example": "f1eebc99-9c0b-4ef8-bb6d-6bb9bd380a22"
          },
          "status": {
            "type": "string",
            "description": "Status atual da afiliação da família.",
            "enum": [
              "PENDING_PAYMENT",
              "NOT_AFFILIATED",
              "AFFILIATED",
              "EXPIRED"
            ],
            "example": "AFFILIATED"
          },
          "affiliatedAt": {
            "type": "string",
            "description": "Data em que a afiliação se tornou ativa.",
            "format": "date-time",
            "example": "2025-07-17T03:00:00.000Z",
            "nullable": true
          },
          "affiliationExpiresAt": {
            "type": "string",
            "description": "Data em que a afiliação irá expirar.",
            "format": "date-time",
            "example": "2026-07-17T03:00:00.000Z",
            "nullable": true
          },
          "dependants": {
            "description": "Lista de dependentes associados a esta família.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DependantDto"
            }
          }
        },
        "required": [
          "id",
          "holderId",
          "status",
          "affiliatedAt",
          "affiliationExpiresAt",
          "dependants"
        ]
      },
      "AffiliationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único da família.",
            "format": "uuid",
            "example": "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11"
          },
          "holder": {
            "description": "Dados do titular.",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserDto"
              }
            ]
          },
          "status": {
            "type": "string",
            "description": "Status atual da afiliação da família.",
            "enum": [
              "PENDING_PAYMENT",
              "NOT_AFFILIATED",
              "AFFILIATED",
              "EXPIRED"
            ],
            "example": "AFFILIATED"
          },
          "affiliatedAt": {
            "type": "string",
            "description": "Data em que a afiliação se tornou ativa.",
            "format": "date-time",
            "example": "2025-07-17T03:00:00.000Z",
            "nullable": true
          },
          "affiliationExpiresAt": {
            "type": "string",
            "description": "Data em que a afiliação irá expirar.",
            "format": "date-time",
            "example": "2026-07-17T03:00:00.000Z",
            "nullable": true
          },
          "dependants": {
            "description": "Lista de dependentes associados a esta família.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DependantDto"
            }
          }
        },
        "required": [
          "id",
          "holder",
          "status",
          "affiliatedAt",
          "affiliationExpiresAt",
          "dependants"
        ]
      },
      "ChangeClubDirectorDto": {
        "type": "object",
        "properties": {
          "newDirectorId": {
            "type": "string",
            "description": "ID do usuário que será o novo diretor do clube.",
            "example": "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11",
            "format": "uuid"
          }
        },
        "required": [
          "newDirectorId"
        ]
      },
      "EnrollmentRequestDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único da solicitação de matrícula.",
            "format": "uuid",
            "example": "a1b2c3d4-e5f6-g7h8-i9j0-k1l2m3n4o5p6"
          },
          "status": {
            "type": "string",
            "description": "Status atual da solicitação.",
            "enum": [
              "REJECTED",
              "APPROVED",
              "PENDING"
            ],
            "example": "PENDING"
          },
          "clubId": {
            "type": "string",
            "description": "ID do clube ao qual a matrícula foi solicitada.",
            "format": "uuid",
            "example": "c1b2c3d4-e5f6-g7h8-i9j0-k1l2m3n4o5p7"
          },
          "familyId": {
            "type": "string",
            "description": "ID da família que está solicitando a matrícula.",
            "format": "uuid",
            "example": "d1b2c3d4-e5f6-g7h8-i9j0-k1l2m3n4o5p8"
          },
          "dependantId": {
            "type": "string",
            "description": "ID do dependente para o qual a matrícula se aplica.",
            "format": "uuid",
            "example": "e1b2c3d4-e5f6-g7h8-i9j0-k1l2m3n4o5p9"
          },
          "requestedAt": {
            "type": "string",
            "description": "Data e hora em que a solicitação foi criada.",
            "format": "date-time",
            "example": "2025-07-20T14:00:00Z"
          },
          "resolvedAt": {
            "type": "string",
            "description": "Data e hora em que a solicitação foi resolvida (aprovada/rejeitada).",
            "format": "date-time",
            "nullable": true,
            "example": "2025-07-20T15:30:00Z"
          },
          "rejectionReason": {
            "type": "object",
            "description": "Motivo da rejeição, caso o status seja REJECTED.",
            "nullable": true,
            "example": "Vagas esgotadas para a faixa etária."
          }
        },
        "required": [
          "id",
          "status",
          "clubId",
          "familyId",
          "dependantId",
          "requestedAt",
          "resolvedAt",
          "rejectionReason"
        ]
      },
      "RegisterUserInputDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "Primeiro nome do responsável familiar.",
            "example": "João",
            "minLength": 2
          },
          "lastName": {
            "type": "string",
            "description": "Último nome do responsável familiar.",
            "example": "Silva",
            "minLength": 2
          },
          "email": {
            "type": "string",
            "description": "Email único para o cadastro.",
            "example": "joao.silva@example.com"
          },
          "phone": {
            "type": "string",
            "description": "Telefone celular do responsável.",
            "example": "11987654321",
            "minLength": 10
          },
          "cpf": {
            "type": "string",
            "description": "CPF do responsável familiar.",
            "example": "12345678900"
          },
          "password": {
            "type": "string",
            "description": "Senha de acesso.",
            "example": "Password@123",
            "minLength": 8
          },
          "confirmPassword": {
            "type": "string",
            "description": "Confirmação da senha de acesso. Deve ser idêntica à senha.",
            "example": "Password@123"
          },
          "address": {
            "description": "Endereço completo do responsável.",
            "allOf": [
              {
                "$ref": "#/components/schemas/AddressDto"
              }
            ]
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "phone",
          "cpf",
          "password",
          "confirmPassword",
          "address"
        ]
      },
...
    }
  }
}